Add UserUpdate and UserDelete services
Introduced UserUpdate and UserDelete services in the application layer. Made changes to UserRepository in the domain layer to facilitate these updates.

These changes improve the modularity of the application and make it easier to handle updates and deletions of user data. Also made some minor adjustments to error handling in UserRepository.