Refactor SQLite queries with new ToSQLite trait
A new ToSQLite trait is introduced which reworks how SQLite queries are generated and executed. This simplifies the construction of complex queries by using the new trait in the UserCriteriaRepositorySQLite to handle filter, order, limit and offset criteria. Leveraging this approach gives better organization to the code and improves readability.